revert the ptl ifl_wrapper changes
